refactor traits into abstract classes
